+++ src/pages/memberSetup/ui/AddMember/index.tsx
@@ -2,10 +2,9 @@ import { useForm } from 'react-hook-form';
 import * as z from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Member } from '@/entities/member/model/member.type';
-import { Button, Input, Profile } from '@/shared/design-system/ui';
+import { Button, Input, Profile, showToast } from '@/shared/design-system/ui';
 import Text from '@/shared/ui/Text';
 import Flex from '@/shared/ui/Flex';
-import { showToast } from '@/shared/ui/Toast';
 import useAddGroupMember from './api/useAddGroupMember';
 import useDeleteGroupMember from './api/useDeleteGroupMember';
 import * as S from './index.styles';
