+++ src/pages/myLinks/ui/LinkCard/LinkCard.tsx
@@ -1,6 +1,6 @@
 import SvgCopy from '@/shared/assets/svgs/icon/Copy';
 import copyClipboard from '@/shared/lib/copyClipboard';
-import { showToast } from '@/shared/ui/Toast';
+import { showToast } from '@/shared/design-system/ui';
 import * as S from './LinkCard.styles';
 
 interface LinkCardProps {
