+++ src/pages/qrStep/QrStepPage.tsx
@@ -4,8 +4,12 @@ import { QRCodeSVG } from 'qrcode.react';
 import { toPng } from 'html-to-image';
 import saveAs from 'file-saver';
 import { ArrowLeft, Reset } from '@/shared/assets/svgs/icon';
-import { showToast } from '@/shared/ui/Toast';
-import { Button, DescriptionField, Header } from '@/shared/design-system/ui';
+import {
+  Button,
+  DescriptionField,
+  Header,
+  showToast,
+} from '@/shared/design-system/ui';
 import { BottomButtonContainer } from '@/shared/styles/bottomButton.styles';
 import generateShareLink from '@/shared/lib/generateShareLink';
 import * as S from './QrStepPage.styles';
