+++ src/shared/design-system/ui/Toast/ToastProvider.tsx
@@ -1,0 +1,6 @@
+import 'react-toastify/dist/ReactToastify.css';
+import * as S from './ToastProvider.styles';
+
+export function ToastProvider() {
+  return <S.StyledToastContainer />;
+}
