+++ src/shared/design-system/ui/Toast/index.ts
@@ -1,2 +1,4 @@
 export { Toast } from './Toast';
 export type { ToastProps, ToastType } from './Toast';
+export { ToastProvider } from './ToastProvider';
+export { showToast } from './showToast';
