--- src/shared/ui/Toast/index.type.ts
@@ -1,6 +1,0 @@
-export type ToastType = 'info' | 'error' | 'success' | 'warning';
-
-export interface ToastProps {
-  type: ToastType;
-  content: string;
-}
